Move parser into ssh.py
