--- docs/shared_parsers_catalog/sshd_test_mode.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.parsers.sshd_test_mode
-   :members:
-   :show-inheritance:
